+++ src/main/java/com/poortorich/chat/manager/ChatroomLeaveManager.java
@@ -4,7 +4,7 @@ import com.poortorich.chat.entity.ChatParticipant;
 import com.poortorich.chat.entity.Chatroom;
 import com.poortorich.chat.entity.enums.ChatroomRole;
 import com.poortorich.chat.realtime.model.PayloadContext;
-import com.poortorich.chat.realtime.payload.BasePayload;
+import com.poortorich.chat.realtime.payload.response.BasePayload;
 import com.poortorich.chat.service.ChatMessageService;
 import com.poortorich.chat.service.ChatParticipantService;
 import com.poortorich.chat.service.ChatroomService;
