+++ src/main/java/com/poortorich/chat/realtime/payload/interfaces/ResponsePayload.java
@@ -1,6 +1,6 @@
 package com.poortorich.chat.realtime.payload.interfaces;
 
-import com.poortorich.chat.realtime.payload.BasePayload;
+import com.poortorich.chat.realtime.payload.response.BasePayload;
 
 public interface ResponsePayload {
